Remove Peoplefinder namespace from validators
