--- app/validators/peoplefinder/email_validator.rb
@@ -1,17 +1,0 @@
-require 'peoplefinder'
-
-module Peoplefinder
-  class EmailValidator < ActiveModel::EachValidator
-    def validate_each(record, attribute, value)
-      email_address = EmailAddress.new(value)
-
-      if email_address.valid_format?
-        unless email_address.valid_domain?
-          record.errors.add(attribute, :invalid_domain, options)
-        end
-      else
-        record.errors.add(attribute, :invalid_format, options)
-      end
-    end
-  end
-end
